% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_methods.R
\docType{methods}
\name{is_ai}
\alias{is_ai}
\alias{is_ai,Game-method}
\alias{set_human}
\alias{set_human,Game-method}
\title{Check if a player is an AI}
\usage{
is_ai(obj, ...)

\S4method{is_ai}{Game}(obj, player, ...)

set_human(obj, ...)

\S4method{set_human}{Game}(obj, player, ...)
}
\arguments{
\item{obj}{A Game object}

\item{...}{Additional objects}

\item{player}{Integer or player name identifying the players who will be played by a human}
}
\description{
Checks if a given player is played by an AI algorithm.

Flag the indicated player as human.
}
