% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jass_classes.R
\docType{class}
\name{Hand}
\alias{Hand}
\alias{Player}
\alias{Trick}
\alias{Round}
\alias{Game}
\title{S4 classes for Jass}
\description{
This contains a description of classes used in the Jass package
}
\section{Slots}{

\describe{
\item{\code{cards}}{A data.frame with 36 rows and the following columns: suit, face, trump (logical), and inhand (logical, indicating whether the given card is in the current hand)}

\item{\code{name}}{Player name (character)}

\item{\code{AI}}{Logical indicating an AI player}

\item{\code{played}}{A list of Jass hands, each containing 1 card. Cards are placed in the list in player order}

\item{\code{lead_suit}}{A string identifying the suit for this trick - can be overridden by Trump}

\item{\code{trump}}{A string listing the suit for this round}

\item{\code{trick}}{A trick object to keep track of cards on the table}

\item{\code{next_player}}{An integer designating which player's turn is next}

\item{\code{hands}}{A list of hands containing hand for each player this round}

\item{\code{teams}}{A vector of team names identifying the team for each player}

\item{\code{won}}{A list of hands containing cards won in previous tricks (one for each player/team)}

\item{\code{history}}{A list of tricks, summing up game play during the current round}

\item{\code{players}}{A list of Jass Player objects, one for each player}

\item{\code{game}}{A character string defining what Jass variant is being played}

\item{\code{round}}{A Round object containing hands}

\item{\code{teams}}{A vector of team names identifying the team for each player}

\item{\code{score}}{An integer vector to keep track of the score for the game}

\item{\code{start}}{An integer indicating which player will start the current round (this is incremented when `deal()` is called on a Game object).}
}}

