% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Jass.R
\name{pick_random_valid_card}
\alias{pick_random_valid_card}
\title{Pick a random, valid card to play}
\usage{
pick_random_valid_card(trk, h, ...)
}
\arguments{
\item{trk}{A Trick object}

\item{h}{A Hand object}

\item{...}{Other arguments that are ignored by this function}
}
\value{
A character value of the card to play
}
\description{
This is simplest function to pick a card to play. It picks a random following the basic Jass rules.
}
