% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_methods.R
\name{cards}
\alias{cards}
\alias{cards<-}
\alias{trump<-}
\alias{trump}
\title{Display cards in a Jass Hand}
\usage{
cards(obj, ...)

cards(x, ...) <- value

trump(x) <- value

trump(obj, ...)
}
\arguments{
\item{obj}{An object of the proper calss}

\item{...}{Other values used in specific methods (see Description)}

\item{x}{An object of the proper class}

\item{value}{A character string containing the trump suit}
}
\description{
This generic will display a hand of Jass cards

This generic will add cards to or discard cards from a hand of Jass cards

Option arguments include: `draw`, which by default adds cards to the hand, but for hand objects causes the specified cards to be discarded when FALSE.

This generic sets Trump for the object and any hands contained in the object

This generic gets Trump for the object and any hands contained in the object
}
