% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_methods.R
\docType{methods}
\name{cards}
\alias{cards}
\alias{card_abbr}
\alias{suitTranslation}
\alias{card_order}
\alias{card_value}
\alias{cards,Hand-method}
\alias{cards,Trick-method}
\alias{cards,Round-method}
\alias{cards,Game-method}
\alias{cards<-}
\alias{cards<-,Hand-method}
\alias{trump<-}
\alias{trump<-,Hand-method}
\alias{trump<-,Round-method}
\alias{trump<-,Game-method}
\alias{trump}
\alias{trump,Hand-method}
\title{card_abbr}
\usage{
card_abbr(cards)

suitTranslation(suit)

card_order(face, trump, game = "Cross Jass")

card_value(face, trump, game = "Cross Jass")

cards(obj, ...)

\S4method{cards}{Hand}(obj, lead_suit = NULL, ...)

\S4method{cards}{Trick}(obj, ...)

\S4method{cards}{Round}(obj, player = 1, ...)

\S4method{cards}{Game}(obj, player = 1, ...)

cards(x, ...) <- value

\S4method{cards}{Hand}(x, draw = TRUE, ...) <- value

trump(x) <- value

\S4method{trump}{Hand}(x) <- value

\S4method{trump}{Round}(x) <- value

\S4method{trump}{Game}(x) <- value

trump(obj, ...)

\S4method{trump}{Hand}(obj, ...)
}
\arguments{
\item{cards}{The cards slot from a hand object (or a subset)}

\item{suit}{Character string naming suit or suit abbreviation}

\item{face}{Character (or character vector) indicating the face to order}

\item{trump}{logical indicating that trump ordering should be used}

\item{game}{Character indicating the game being played. Most result in identical ordering.}

\item{obj}{An object of the proper class}

\item{...}{Other values used in specific methods (see Description)}

\item{lead_suit}{A character defining the suit that was lead in the current trick (ignored when NULL)}

\item{player}{An integer indicating the player who's hand should be displayed (ignored when not needed)}

\item{x}{An object of the proper class}

\item{value}{A character string containing the trump suit}

\item{draw}{Logical - adds cards to the hand when TRUE, but for hand objects causes the specified cards to be discarded by setting `draw` equal to FALSE.}
}
\value{
A string of abbreviated card names



Returns a number (or numeric vector) indicating card order (1 being first, 9 being last)

Returns a number (or numeric vector) indicating card point value(s)
}
\description{
Abbreviate cards names in a hand.

Translate between suit names and abbreviations

Return the order of cards for trump/non-trump suits

Return the point value of cards for trump/non-trump suits

This generic will display a hand of Jass cards

This generic will add cards to or discard cards from a hand of Jass cards

This generic sets Trump for the object and any hands contained in the object

This generic gets Trump for the object and any hands contained in the object
}
