% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jass-classes.R
\docType{class}
\name{Hand}
\alias{Hand}
\alias{Trick}
\title{S4 classes for Jass}
\description{
This contains a description of classes used in the Jass package
}
\section{Slots}{

\describe{
\item{\code{cards}}{A data.frame with 36 rows and the following columns: suit, face, trump (logical), and inhand (logical, indicating whether the given card is in the current hand)}

\item{\code{hands}}{A list of Jass hands, each containing 1 card. Cards are placed in the list in the order they are played.}
}}

