% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_methods.R
\name{deal}
\alias{deal}
\alias{play}
\title{Deal a round of cards}
\usage{
deal(obj, game = NULL, n = NULL, ...)

play(obj, rules, ...)
}
\arguments{
\item{obj}{A Game object}

\item{game}{A character value defining the game being played}

\item{n}{A numeric value defining the number of players or hands to be dealt}

\item{...}{Other arguments}

\item{rules}{A function to decide which card to play}
}
\description{
Deals a round of cards.

This generic plays a valid card based on the current state of the game
}
