% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_methods.R
\docType{methods}
\name{deal}
\alias{deal}
\alias{deal,Round-method}
\alias{deal,Game-method}
\alias{play}
\alias{play,Game-method}
\alias{status}
\alias{status,Round-method}
\alias{status,Game-method}
\alias{next_trick}
\alias{next_trick,Round-method}
\alias{next_trick,Game-method}
\title{Deal a round of cards}
\usage{
deal(obj, game = NULL, n = NULL, ...)

\S4method{deal}{Round}(obj, game = "Cross Jass", n = 4, ...)

\S4method{deal}{Game}(obj, game = "Cross Jass", n = 4, ...)

play(obj, ...)

\S4method{play}{Game}(obj, to_play = NULL, rules = pick_random_valid_card, state = NULL, ...)

status(obj, verbose = TRUE, ...)

\S4method{status}{Round}(obj, verbose = TRUE, ...)

\S4method{status}{Game}(obj, verbose = TRUE, ...)

next_trick(obj, ...)

\S4method{next_trick}{Round}(obj, verbose = TRUE, ...)

\S4method{next_trick}{Game}(obj, verbose = TRUE, ...)
}
\arguments{
\item{obj}{An object of the correct class}

\item{game}{A character value defining the game being played}

\item{n}{A numeric value defining the number of players or hands to be dealt}

\item{...}{Other arguments for specific classes}

\item{to_play}{Character string identifying the card to play (for non-AI players)}

\item{rules}{A function to decide which card to play (for AI players)}

\item{state}{An object containing information about the game, used as input for the model defined by `rules`}

\item{verbose}{Logical indicating that we should print out status results}
}
\description{
Deals a round of cards.

This generic plays a valid card based on a model provided and the current state of the game

This generic returns a list containing the current game status

This generic advances the round to the next trick
}
